refactoring: correct singleton
